refactor(native): split chara_detail_scene_context into .h/.cpp
Fourth step of the header/.cpp split, applied to the scene context (235 lines).
This header is included by other headers (recognizer, scene_scraper,
frame_distributor), so its public surface stays in the header while the
implementation moves out.

The header keeps:
- The public types other headers depend on: TabPage, tabPageTag, kAllTabPages,
  TabCondition, SceneInfo, SceneState.
- CharaDetailSceneContext's two private function templates (resolveBranches,
  firstMet), which must stay in the header, plus the class declaration and
  members.

The non-template method bodies (ctor, update, met, onIdle, the scene begin/end
debounce helpers, getActiveTabIndex/getRecordType) move to the new .cpp verbatim.

Both consuming CMake source lists are updated in lockstep (native/CMakeLists.txt
and windows/runner/CMakeLists.txt).

Verified: umacapture_cli builds, umacapture_tests still passes (12/22),
`flutter build windows --debug` links the runner, and a full-pipeline video smoke
(sandbox/inheritance_only_1.mp4, temporarily using the horizontal crop path)
exercised the scene begin/update/end path end to end -- onCharaDetailStarted,
onPageReady/onScrollReady, onCharaDetailClosed, onCharaDetailFinished success:true.
