+++ native/src/chara_detail/chara_detail_scene_context.h
@@ -59,109 +59,31 @@ public:
         const event_util::Sender<Frame, SceneState> &on_scene_updated,
         const event_util::Sender<> &on_scene_end,
         const chrono_util::time_unit &scene_begin_timeout,
-        const chrono_util::time_unit &scene_end_timeout)
-        : child(child)
-        , tab_page_condition(dynamic_cast<const TabCondition *>(child->findByTag("tab_page")))
-        , record_type_condition(dynamic_cast<const TabCondition *>(child->findByTag("record_type")))
-        , on_scene_begin(on_scene_begin)
-        , on_scene_updated(on_scene_updated)
-        , on_scene_end(on_scene_end)
-        , scene_begin_timeout(scene_begin_timeout)
-        , scene_end_timeout(scene_end_timeout) {
-        if (tab_page_condition == nullptr) {
-            throw std::runtime_error("tab_page condition not found");
-        }
-        if (record_type_condition == nullptr) {
-            throw std::runtime_error("record_type condition not found");
-        }
-        record_type_branches = resolveBranches(record_type_condition, record::kAllRecordTypes, record::recordTypeTag);
-        tab_page_branches = resolveBranches(tab_page_condition, kAllTabPages, tabPageTag);
-    }
+        const chrono_util::time_unit &scene_end_timeout);
 
-    void update(const Frame &input) override {
-        child->update(input);
-        const auto tab_page = getActiveTabIndex();
-        const auto record_type = getRecordType();
-        met_ = child->met() && tab_page.has_value() && record_type.has_value();
+    void update(const Frame &input) override;
 
-        if (met_) {
-            scene_end_pending_since = std::nullopt;  // A reappearance within the timeout keeps the same scene.
-            if (!scene_active) {
-                beginSceneWhenStable(input, record_type.value());
-            }
-            // Scrape only after the scene has actually begun. Holding back until the begin debounce commits
-            // also makes the scraper's reference frame a settled, post-animation still instead of one captured
-            // mid-animation.
-            if (scene_active) {
-                on_scene_updated->send(input, {tab_page.value(), record_type.value()});
-            }
-        } else {
-            // Any drop before commit resets the begin window; the same record_type must persist uninterrupted.
-            scene_begin_pending_since = std::nullopt;
-            scene_begin_pending_type = std::nullopt;
-            if (scene_active) {
-                // Debounce the scene end by video time (frame timestamps), not wall-clock. A wall-clock timer
-                // running on its own thread made scene closing depend on how fast frames were fed during
-                // offline video replay; keying off the frame timestamp keeps this deterministic. Frame
-                // timestamps track real time in live capture, so live behavior is unchanged.
-                if (scene_end_timeout == chrono_util::time_unit::zero()) {
-                    endScene();
-                } else if (!scene_end_pending_since) {
-                    scene_end_pending_since = input.timestamp();
-                } else if (chrono_util::monotonicElapsed(input.timestamp(), scene_end_pending_since.value())
-                           >= sceneEndTimeoutMs()) {
-                    endScene();
-                }
-            }
-        }
-    }
-
-    [[nodiscard]] bool met() const override { return met_; }
+    [[nodiscard]] bool met() const override;
 
     // The live frame stream stalled. The frame-timestamp scene-end debounce cannot advance without frames, so
     // close an already-committed scene here; a scene still in the begin debounce never committed, so just drop
     // its pending window. Invoked on the distributor runner thread, the same thread as update().
-    void onIdle() override {
-        if (scene_active) {
-            endScene();
-        }
-        scene_begin_pending_since = std::nullopt;
-        scene_begin_pending_type = std::nullopt;
-    }
+    void onIdle() override;
 
 private:
     // Commit the scene only once the same record_type has stayed met for the begin timeout. The record_type is
     // locked here for the whole scene, so a transient first-match during the opening animation must not win — it
     // has to persist. Keyed off frame timestamps for replay-deterministic timing, mirroring the scene-end
     // debounce. A zero timeout commits on the first met frame (legacy behavior).
-    void beginSceneWhenStable(const Frame &input, record::RecordType record_type) {
-        if (scene_begin_timeout == chrono_util::time_unit::zero()) {
-            beginScene(record_type);
-        } else if (!scene_begin_pending_since || scene_begin_pending_type != record_type) {
-            scene_begin_pending_since = input.timestamp();
-            scene_begin_pending_type = record_type;
-        } else if (chrono_util::monotonicElapsed(input.timestamp(), scene_begin_pending_since.value())
-                   >= sceneBeginTimeoutMs()) {
-            beginScene(record_type);
-        }
-    }
+    void beginSceneWhenStable(const Frame &input, record::RecordType record_type);
 
-    void beginScene(record::RecordType record_type) {
-        on_scene_begin->send({record_type});
-        scene_active = true;
-        scene_begin_pending_since = std::nullopt;
-        scene_begin_pending_type = std::nullopt;
-    }
+    void beginScene(record::RecordType record_type);
 
-    void endScene() {
-        on_scene_end->send();
-        scene_active = false;
-        scene_end_pending_since = std::nullopt;
-    }
+    void endScene();
 
-    [[nodiscard]] uint64 sceneBeginTimeoutMs() const { return static_cast<uint64>(scene_begin_timeout.count()); }
+    [[nodiscard]] uint64 sceneBeginTimeoutMs() const;
 
-    [[nodiscard]] uint64 sceneEndTimeoutMs() const { return static_cast<uint64>(scene_end_timeout.count()); }
+    [[nodiscard]] uint64 sceneEndTimeoutMs() const;
 
     // Bind each enum value to its named branch in `cond` once, at construction. Resolving by tag (not by
     // child position) means reordering the branch list cannot silently remap a value, and a missing or
@@ -205,13 +127,9 @@ private:
         return found;
     }
 
-    [[nodiscard]] std::optional<TabPage> getActiveTabIndex() const {
-        return firstMet(tab_page_branches, kAllTabPages);
-    }
+    [[nodiscard]] std::optional<TabPage> getActiveTabIndex() const;
 
-    [[nodiscard]] std::optional<record::RecordType> getRecordType() const {
-        return firstMet(record_type_branches, record::kAllRecordTypes);
-    }
+    [[nodiscard]] std::optional<record::RecordType> getRecordType() const;
 
     const std::shared_ptr<condition::Condition<Frame>> child;
     const condition::ParallelCondition<Frame, rule::LogicalOr> *tab_page_condition;
